fix apply !override while processing extends on services declared in same file
