+++ loader/loader.go
@@ -427,7 +427,7 @@ func loadYamlFile(ctx context.Context,
 		file.Content = content
 	}
 
-	processRawYaml := func(raw interface{}, processors ...PostProcessor) error {
+	processRawYaml := func(raw interface{}, processor PostProcessor) error {
 		converted, err := convertToStringKeysRecursive(raw, "")
 		if err != nil {
 			return err
@@ -447,16 +447,14 @@ func loadYamlFile(ctx context.Context,
 		fixEmptyNotNull(cfg)
 
 		if !opts.SkipExtends {
-			err = ApplyExtends(ctx, cfg, opts, ct, processors...)
+			err = ApplyExtends(ctx, cfg, opts, ct, processor)
 			if err != nil {
 				return err
 			}
 		}
 
-		for _, processor := range processors {
-			if err := processor.Apply(dict); err != nil {
-				return err
-			}
+		if err := processor.Apply(dict); err != nil {
+			return err
 		}
 
 		if !opts.SkipInclude {
@@ -519,7 +517,7 @@ func loadYamlFile(ctx context.Context,
 			}
 		}
 	} else {
-		if err := processRawYaml(file.Config); err != nil {
+		if err := processRawYaml(file.Config, NoopPostProcessor{}); err != nil {
 			return nil, nil, err
 		}
 	}
